💄 SpeedDialButton の代わりにドロワーメニューボタンを使用するよう変更
